Enhanced command pattern and refactoring
- Removed `ComponentChanged.h` from `CommandBuffer.h` includes, added `assert.h`.
- Updated `CommandBuffer` to use static cast for `cursor` assignment.
- Modified `ComponentChanged.h` include path and enhanced `Merge` function.
- Introduced `VarChanged` class for advanced undo/redo functionality.
- Updated project files to include `VarChanged` class, ensuring compilation.
- Reorganized project structure for better readability and maintenance.

+++ Source/Editor/Editor/Commands/CommandBuffer.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <vector>
-#include <Editor/Editor/Commands/ComponentChanged.h>
+#include <assert.h> 
 
 class BaseCommand
 {
@@ -41,7 +41,7 @@ public:
 
 		commandList.erase(commandList.begin() + cursor + 1, commandList.end());
 		commandList.push_back(ptr);
-		cursor = commandList.size() - 1;
+		cursor = static_cast<int>(commandList.size()) - 1;
 	};
